+++ tests/test_vector_quantize_pytorch.py
@@ -11,13 +11,15 @@ class TestVectorQuantizer:
 
     quantizer = VectorQuantize(
         dim=dim,
-        codebook_size=codebook_size,  # codebook size
         commitment_weight=1.0,  # the weight on the commitment loss
         codebook_params=codebook_params,
     )
 
     def test_init(self):
         assert self.quantizer
+
+    def test_not_use_cosine(self):
+        assert not self.quantizer._codebook.use_cosine_sim
 
     def test_forward(self, vectors_channel_last):
         vectors = vectors_channel_last
@@ -36,7 +38,6 @@ class TestVectorQuantizerChannelFirst:
 
     quantizer = VectorQuantize(
         dim=dim,
-        codebook_size=codebook_size,  # codebook size
         commitment_weight=1.0,  # the weight on the commitment loss
         channel_last=False,
         codebook_params=codebook_params,
@@ -58,16 +59,19 @@ class TestVectorQuantizerChannelFirst:
 class TestVectorQuantizerCosine:
     dim = 4
     codebook_size = 2**5
-    codebook_params = CodebookParams(dim=dim, codebook_size=codebook_size)
+    codebook_params = CodebookParams(
+        dim=dim, codebook_size=codebook_size, use_cosine_sim=True
+    )
     quantizer = VectorQuantize(
         dim=4,
-        codebook_size=2**5,  # codebook size
-        use_cosine_sim=True,
         codebook_params=codebook_params,
     )
 
     def test_init(self):
         assert self.quantizer
+
+    def test_use_cosine(self):
+        assert self.quantizer._codebook.use_cosine_sim
 
     def test_forward(self, get_vectors):
         vectors = get_vectors(dim=self.dim)
@@ -90,7 +94,6 @@ class TestVectorQuantizerMultihead:
         codebook_dim=codebook_dim,  # a number of papers have shown smaller codebook dimension to be acceptable
         heads=heads,  # number of heads to vector quantize, codebook shared across all heads
         separate_codebook_per_head=True,  # whether to have a separate codebook per head. False would mean 1 shared codebook
-        codebook_size=2**5,
         codebook_params=codebook_params,
     )
 
@@ -123,7 +126,6 @@ class TestVectorQuantizerMultiheadWithKmeansInit:
         codebook_dim=codebook_dim,  # a number of papers have shown smaller codebook dimension to be acceptable
         heads=heads,  # number of heads to vector quantize, codebook shared across all heads
         separate_codebook_per_head=True,  # whether to have a separate codebook per head. False would mean 1 shared codebook
-        codebook_size=2**5,
         codebook_params=codebook_params,
     )
 
@@ -146,7 +148,6 @@ class TestVectorQuantizerLowerCode:
     codebook_params = CodebookParams(dim=dim, codebook_size=codebook_size)
     quantizer = VectorQuantize(
         dim=4,
-        codebook_size=256,
         codebook_dim=2,  # paper proposes setting this to 32 or as low as 8 to increase codebook usage
         codebook_params=codebook_params,
     )
@@ -177,7 +178,6 @@ class TestVectorQuantizerKmeansInit:
 
     quantizer = VectorQuantize(
         dim=4,
-        codebook_size=codebook_size,
         commitment_weight=1.0,
         codebook_params=codebook_params,
     )
@@ -204,14 +204,13 @@ class TestVectorQuantizerKmeansInitWithCosine:
         codebook_size=2**5,
         initialization_by_kmeans=True,
         kmeans_params=kmeans_params,
+        use_cosine_sim=True,
     )
 
     quantizer = VectorQuantize(
         dim=4,
-        codebook_size=codebook_size,
         commitment_weight=1.0,
         codebook_params=codebook_params,
-        use_cosine_sim=True,
     )
 
     def test_init(self):
@@ -240,7 +239,6 @@ class TestVectorQuantizerKmeansInitWithFewSamples:
 
     quantizer = VectorQuantize(
         dim=4,
-        codebook_size=codebook_size,
         commitment_weight=1.0,
         codebook_params=codebook_params,
     )