+++ Changelog.md
@@ -9,6 +9,8 @@
 * Transform gumbel_params and kmeans_params into dict if they are still dataclasses at initialization. This avoid error if the kmeans_params and gumbel_params come from 'asdict'-ification of CodebookParams early on.
 * CHange the tests accordingly.
 * Rename the method 'replace' in Codebook into 'replace_codes' to avoid confusion with 'replace' from Dataclasses library.
+* The CodebookParams contain now everything needed to instanciate a codebook. For instance, no more GumbelParams in VQ.
+* Changes made accordingly in the ResidualVQ API and test suit.
 
 
 ## [Version 1.16.3]